(PC-20414) refactor(AlgoliaVenue): move type to libs/algolia/types.ts
